Change the sequencer register with address 1 from 0x00 to 0x01.
This bit corresponds to seq_8dot_char bit. If this bit is 0, dot_cnt counts
from 0 to 8, so each horiz_cnt is 9 clock cycles. As horiz_cnt goes from 0 to
63h it does 100 counts, 900 clock cycles in total. That corresponds to a hsync
frequency of 28.89 kHz, and vsync of 64.63 Hz (this is out of range for my
monitor to sync).

However in the OSD menu, seq_8dot_char is 1, dot_cnt counts from 0 to 7, so
horiz_cnt has 800 clock cycles, with a hsync frequency of 32.5 kHz and vsync of
72.71 Hz. This is correct to my monitor.